+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJZ6J3E4FC4R82V464V7GQXT"
+        "01KK0WE94FSXKVXD9JSQC8K73Q"
     )
     r"""The unique identifier of the data source"""