+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJZ6J3E0GKAJ2PQD1ZCXG8E4"
+        "01KK0WE944Y9DN8JKS279Z16E8"
     )
     r"""The unique identifier of the data source"""