+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJZ6J3E3B6BV4J6NCKSRFY20"
+        "01KK0WE94C6ENNEM5RATXAPH9N"
     )
     r"""The unique identifier of the data source"""