+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJZ6J3E8X7SHXAFDJWPN9EBG"
+        "01KK0WE94WCRFA8R9AGX75K0PX"
     )
     r"""The unique identifier of the data source"""