+++ packages/orq-rc/RELEASES.md
@@ -6319,3 +6319,13 @@ Based on:
 - [python v4.5.0-rc.33] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.33] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.33 - packages/orq-rc
+
+## 2026-03-06 06:15:25
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.745.0 (2.850.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.35] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.35] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.35 - packages/orq-rc